refactor(install): move ConfigValidator interface closer to its ussage
